Simplify file structure for the Alpaka port
